👡🌼 Checkpoint ./komutlar/botekle-onayla.js:23205691/3330 ./bot.js:23205691/6496

+++ komutlar/botekle-onayla.js
@@ -3,7 +3,31 @@ const ayarlar = require("../ayarlar.json");
 const db = require('quick.db')
 exports.run = function(client, message, args) {
   let yetkiliROL = ayarlar.yetkiliROL;
-  if (!message.member.roles.cache.has(yetkiliROL)) return message.channel.send(hata1)
+  let hata1 = new Discord.MessageEmbed()
+  .setDescription(`
+   ・ **Hata 01**
+   
+   \`Bu komutu sadece\` <@&${yetkiliROL}> \`rolüne sahip olanlar kullanabilir.\`
+`)
+    let hata2 = new Discord.MessageEmbed()
+  .setDescription(`
+   ・ **Hata 02**
+   
+   \`Onaylanacak botun sahibini sunucuda bulamıyorum.\`
+`)
+       let hata3 = new Discord.MessageEmbed()
+  .setDescription(`
+   ・ **Hata 03**
+   
+   \`Onaylanacak botu sunucuda bulamıyorum.\`
+`)
+            let hata4 = new Discord.MessageEmbed()
+  .setDescription(`
+   ・ **Hata 04**
+   
+   \`Onaylayacağın botun prefixini belirt.\`
+`)
+ 
   let yetkili = message.author;
   let sahip = message.guild.members.cache.get(args[0]);
   let botisim = message.guild.members.cache.get(args[1]);
@@ -12,31 +36,12 @@ exports.run = function(client, message, args) {
   let log = ayarlar.log;
   let rol = ayarlar.developerROL;
   let b = ayarlar.botROL;
-   let hata1 = new Discord.MessageEmbed()
-  .setDescription(`
-   Hata 01
-   
-   \`Bu komutu sadece <@&${yetkiliROL}> rolüne sahip olanlar kullanabilir.\`
-`)
-    let hata2 = new Discord.MessageEmbed()
-  .setDescription(`
-   Hata 02
-   
-   \`Onaylanacak botun sahibini sunucuda bulamıyorum.\`
-`)
-       let hata3 = new Discord.MessageEmbed()
-  .setDescription(`
-   Hata 03
-   
-   \`Onaylanacak botu sunucuda bulamıyorum.\`
-`)
-            let hata4 = new Discord.MessageEmbed()
-  .setDescription(`
-   Hata 04
-   
-   \`Onaylayacağın botun prefixini belirt.\`
-`)
-  message.delete();
+  
+  if(!message.member.roles.cache.has(yetkiliROL)) return message.channel.send(hata1);
+  if(!sahip) return message.channel.send(hata2);
+  if(!botisim) return message.channel.send(hata3);
+  if(!prefix) return message.channel.send(hata4);
+  message.delete() 
   sahip.roles.add(rol)
   
     botisim.roles.add(b)
